requirements.txt
matplotlib
imgaug
distinctipy
numpy
tensorflow
pillow
plantcv
rembg
shutils
flake8
scikit-learn